requirements.txt
# Django и основные зависимости
Django==4.2.7
djangorestframework==3.14.0
django-environ==0.11.2
django-cors-headers==4.3.1

# JWT аутентификация
djangorestframework-simplejwt==5.3.0

# База данных
PyMySQL==1.1.0

# API документация
drf-yasg==1.21.7

# Фильтрация и пагинация
django-filter==23.3
djangorestframework-filters==1.0.0.dev2

# Кэширование
redis==5.0.1
django-redis==5.4.0

# Валидация и сериализация
marshmallow==3.20.1
marshmallow-dataclass==8.6.0

# Утилиты
python-decouple==3.8
celery==5.3.4
kombu==5.3.4

# Тестирование
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2

# Разработка
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-celery-beat

# Безопасность
django-ratelimit==4.1.0
cryptography==41.0.7

# Логирование
structlog==23.2.0

# Мониторинг
sentry-sdk==1.38.0

# Дополнительные утилиты
Pillow==10.1.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
pandas
reportlab
flower
gunicorn
gevent
setuptools